Add files via upload
Run the .exe. The .pyw is the source code. It borrows one function and a variable from scriptforgui.py - I'll include here as well.

Change the value in the 'Press key(s):" textbox to change what button it presses to keep you online. Currently does not support arrow keys as they can't be entered in text boxes.

Change the value in the "Repeat every x seconds:" textbox to make it activate more or less frequently.

Press the button that says "I'm back." or "I want to afk." to toggle it on/off.